chore: refactor lib directory
