+++ components.json
@@ -5,7 +5,7 @@
   "tsx": true,
   "tailwind": {
     "config": "tailwind.config.js",
-    "css": "lib/styles/index.css",
+    "css": "lib/styles.css",
     "baseColor": "zinc",
     "cssVariables": false
   },
