+++ lib/utils.ts
@@ -1,6 +1,8 @@
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
 import type { User } from "@/db/schema";
+import { getToken } from "next-auth/jwt";
+import { NextRequest } from "next/server";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -22,3 +24,40 @@ export async function getLocalUser() {
 export async function removeLocalUser() {
   await localStorage.removeItem("user");
 }
+
+export const decodeToken = async (req: NextRequest) => {
+  const token = await getToken({ req });
+  if (!token) {
+    return null;
+  }
+  return {
+    id: token.sub,
+    name: token.name,
+    email: token.email,
+    accountId: token.accountId,
+  };
+};
+
+export const filterUserData = async (data: User) => {
+  return {
+    name: data.name,
+    email: data.email,
+    profileImage: data.profileImage,
+    showDocuments: data.showDocuments,
+    showBooks: data.showBooks,
+  };
+};
+
+export const filterOutPassword = (array: Array<User>) => {
+  const filtered = array.map((data: User) => {
+    return {
+      id: data.id,
+      name: data.name,
+      email: data.email,
+      accountId: data.accountId,
+      createdAt: data.createdAt,
+      updatedAt: data.updatedAt,
+    };
+  });
+  return filtered;
+};
